requirements.txt
flask
python-dotenv
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
email-validator
flask-bootstrap
flask-babel
pyjwt